Initial support for caching results, with LZ4
Quite a messy commit, hopefully new ones should be shorter

+++ test/test_c.c
@@ -82,7 +82,7 @@ bool decode (uint32_t mysize, float drop_prob, uint8_t overhead)
 	 * if needed you'll just block on the call.
 	 */
 
-	if (drop_prob > 90.0)
+	if (drop_prob > (float)90.0)
 		drop_prob = 90.0;	// this is still too high probably.
 
 
@@ -171,7 +171,7 @@ bool decode (uint32_t mysize, float drop_prob, uint8_t overhead)
 		}
 		if (sym_rep == RaptorQ_max_repair (enc, block)) {
 			fprintf(stderr, "Maybe losing %f%% symbols is too much?\n",
-																	drop_prob);
+															(double)drop_prob);
 			free (myvec);
 			for (uint32_t k = 0; k < next_encoded; ++k)
 				free (encoded[k].symbol);
@@ -269,9 +269,9 @@ bool decode (uint32_t mysize, float drop_prob, uint8_t overhead)
 	for (uint16_t i = 0; i < mysize; ++i) {
 		if (myvec[i] != received[i]) {
 			fprintf(stderr, "FAILED, but we though otherwise! %i - %f: %i "
-													"%i -- %i\n",
-													mysize, drop_prob, i,
-													myvec[i], received[i]);
+												"%i -- %i\n",
+												mysize, (double) drop_prob, i,
+												myvec[i], received[i]);
 			free (myvec);
 			free (received);
 			for (uint32_t k = 0; k < next_encoded; ++k)
